[refactor] 거리 타입 변환 및 허브 응답 dto수정

+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/domain/vo/DeliveryActual.java
@@ -11,11 +11,11 @@ import lombok.NoArgsConstructor;
 @Embeddable
 public class DeliveryActual {
 	@Column(name = "actual_distance")
-	private Long distance;
+	private Integer distance;
 	@Column(name = "actual_duration")
 	private Integer duration;
 
-	public static DeliveryActual of(Long distance, Integer duration) {
+	public static DeliveryActual of(Integer distance, Integer duration) {
 		if (distance < 0) {
 			throw new IllegalArgumentException("거리는 0 이상이어야 합니다. 단위: 미터");
 		}
