+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/presentation/dto/ProcessHubShippingRequest.java
@@ -6,7 +6,7 @@ public record ProcessHubShippingRequest(
 	UUID deliveryId,
 	UUID deliveryRouteId,
 	Boolean isDeparture,
-	Long distance,
+	Integer distance,
 	Integer duration
 ) {
